@@ -6,62 +6,88 @@ build-backend = "setuptools.build_meta"
 name = "atomate2"
 description = "atomate2 is a library of materials science workflows"
 readme = "README.md"
-keywords = ["high-throughput", "automated", "workflow", "dft", "vasp"]
+keywords = ["automated", "dft", "high-throughput", "vasp", "workflow"]
 license = { text = "modified BSD" }
 authors = [{ name = "Alex Ganose", email = "alexganose@gmail.com" }]
 dynamic = ["version"]
 classifiers = [
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
     "Development Status :: 5 - Production/Stable",
+    "Intended Audience :: Information Technology",
     "Intended Audience :: Science/Research",
     "Intended Audience :: System Administrators",
-    "Intended Audience :: Information Technology",
     "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.9",
     "Topic :: Other/Nonlisted Topic",
     "Topic :: Scientific/Engineering",
 ]
-requires-python = '>="3.8"'
-dependencies =[
-    "pymatgen>=2022.8.23",
-    "custodian>=2022.5.26",
-    "pydantic",
-    "monty",
-    "jobflow>=0.1.5",
+requires-python = ">=3.9"
+dependencies = [
     "PyYAML",
-    "numpy",
     "click",
+    "custodian>=2023.3.10",
+    "emmet-core>=0.70.5",
+    "jobflow>=0.1.11",
+    "monty>=2023.9.25",
+    "numpy",
+    "pydantic-settings>=2.0.3",
+    "pydantic>=2.0.1",
+    "pymatgen>=2023.10.11",
 ]
 
 [project.optional-dependencies]
 amset = ["amset>=0.4.15", "pydash"]
 cclib = ["cclib"]
-mp = ["mp-api>=0.27.5"]
+mp = ["mp-api>=0.37.5"]
 phonons = ["phonopy>=1.10.8", "seekpath"]
+lobster = ["ijson>=3.2.2", "lobsterpy>=0.3.2"]
+defects = ["dscribe>=1.2.0", "pymatgen-analysis-defects>=2022.11.30"]
+forcefields = [
+    "chgnet>=0.2.2",
+    "mace@git+https://github.com/ACEsuit/mace@develop",
+    "matgl>=0.9.0",
+    "quippy-ase>=0.9.14",
+]
 docs = [
-    "numpydoc==1.4.0",
-    "ipython==8.5.0",
     "FireWorks==2.0.3",
-    "autodoc_pydantic==1.7.2",
-    "jupyter-book==0.13.1",
+    "autodoc_pydantic==2.0.1",
+    "furo==2023.9.10",
+    "ipython==8.18.0",
+    "jsonschema[format]",
+    "myst_parser==2.0.0",
+    "numpydoc==1.6.0",
+    "sphinx-copybutton==0.5.2",
+    "sphinx==7.2.6",
+    "sphinx_design==0.5.0",
 ]
 dev = ["pre-commit>=2.12.1"]
-tests = ["pytest==7.1.3", "pytest-cov==4.0.0", "FireWorks==2.0.3"]
+tests = ["FireWorks==2.0.3", "pytest-cov==4.1.0", "pytest==7.4.3"]
 strict = [
-    "pydantic==1.10.2",
-    "pymatgen==2022.9.21",
-    "custodian==2022.5.26",
-    "monty==2022.9.9",
-    "jobflow==0.1.9",
-    "click==8.1.3",
-    "PyYAML==6.0",
-    "cclib==1.7.2",
-    "phonopy==2.16.2",
-    "seekpath==2.0.1",
+    "PyYAML==6.0.1",
+    "cclib==1.8",
+    "chgnet==0.3.2",
+    "click==8.1.7",
+    "custodian==2023.10.9",
+    "dscribe==2.1.0",
+    "emmet-core==0.74.2",
+    "ijson==3.2.3",
+    "jobflow==0.1.14",
+    "lobsterpy==0.3.2",
+    "mace@git+https://github.com/ACEsuit/mace@develop",
+    "matgl==0.9.1",
+    "monty==2023.9.25",
+    "mp-api==0.37.5",
     "numpy",
-    "mp-api==0.27.5"
+    "phonopy==2.20.0",
+    "pydantic-settings==2.1.0",
+    "pydantic==2.4.2",
+    "pymatgen-analysis-defects==2023.10.19",
+    "pymatgen==2023.11.12",
+    "quippy-ase==0.9.14",
+    "seekpath==2.1.0",
+    "typing-extensions==4.8.0",
 ]
 
 [project.scripts]
