@@ -1,13 +1,15 @@
 """Settings for atomate2."""
 
-from pathlib import Path
-from typing import Optional, Tuple, Union
+from __future__ import annotations
 
-from pydantic import BaseSettings, Field, root_validator
+import warnings
+from pathlib import Path
+from typing import Literal, Optional, Union
+
+from pydantic import Field, model_validator
+from pydantic_settings import BaseSettings, SettingsConfigDict
 
 _DEFAULT_CONFIG_FILE_PATH = "~/.atomate2.yaml"
-
-__all__ = ["Atomate2Settings"]
 
 
 class Atomate2Settings(BaseSettings):
@@ -30,7 +32,12 @@ class Atomate2Settings(BaseSettings):
     SYMPREC: float = Field(
         0.1, description="Symmetry precision for spglib symmetry finding."
     )
-    CUSTODIAN_SCRATCH_DIR: str = Field(
+    BANDGAP_TOL: float = Field(
+        1e-4,
+        description="Tolerance for determining if a material is a semiconductor or "
+        "metal",
+    )
+    CUSTODIAN_SCRATCH_DIR: Optional[str] = Field(
         None, description="Path to scratch directory used by custodian."
     )
 
@@ -44,22 +51,19 @@ class Atomate2Settings(BaseSettings):
     VASP_NCL_CMD: str = Field(
         "vasp_ncl", description="Command to run non-collinear version of VASP."
     )
-    VASP_VDW_KERNEL_DIR: str = Field(None, description="Path to VDW VASP kernel.")
+    VASP_VDW_KERNEL_DIR: Optional[str] = Field(
+        None, description="Path to VDW VASP kernel."
+    )
     VASP_INCAR_UPDATES: dict = Field(
         default_factory=dict, description="Updates to apply to VASP INCAR files."
-    )
-    VASP_RELAX_MAX_FORCE: float = Field(
-        0.25,
-        description="Maximum force allowed on each atom for successful structure "
-        "optimization",
     )
     VASP_VOLUME_CHANGE_WARNING_TOL: float = Field(
         0.2,
         description="Maximum volume change allowed in VASP relaxations before the "
         "calculation is tagged with a warning",
     )
-    VASP_HANDLE_UNSUCCESSFUL: Union[str, bool] = Field(
-        "fizzle",
+    VASP_HANDLE_UNSUCCESSFUL: Union[bool, Literal["error"]] = Field(
+        "error",
         description="Three-way toggle on what to do if the job looks OK but is actually"
         " unconverged (either electronic or ionic). - True: mark job as COMPLETED, but "
         "stop children. - False: do nothing, continue with workflow as normal. 'error':"
@@ -68,18 +72,107 @@ class Atomate2Settings(BaseSettings):
     VASP_CUSTODIAN_MAX_ERRORS: int = Field(
         5, description="Maximum number of errors to correct before custodian gives up"
     )
-    VASP_STORE_VOLUMETRIC_DATA: Optional[Tuple[str]] = Field(
+    VASP_STORE_VOLUMETRIC_DATA: Optional[tuple[str]] = Field(
         None, description="Store data from these files in database if present"
     )
     VASP_STORE_ADDITIONAL_JSON: bool = Field(
-        True,
+        default=True,
         description="Ingest any additional JSON data present into database when "
         "parsing VASP directories useful for storing duplicate of FW.json",
     )
     VASP_RUN_BADER: bool = Field(
-        False,
+        default=False,
         description="Whether to run the Bader program when parsing VASP calculations."
         "Requires the bader executable to be on the path.",
+    )
+    VASP_RUN_DDEC6: bool = Field(
+        default=False,
+        description="Whether to run the DDEC6 program when parsing VASP calculations."
+        "Requires the chargemol executable to be on the path.",
+    )
+    DDEC6_ATOMIC_DENSITIES_DIR: Optional[str] = Field(
+        default=None,
+        description="Directory where the atomic densities are stored.",
+        # TODO uncomment below once that functionality is actually implemented
+        # If not set, pymatgen tries to auto-download the densities and extract them
+        # into ~/.cache/pymatgen/ddec
+    )
+
+    VASP_ZIP_FILES: Union[bool, Literal["atomate"]] = Field(
+        "atomate",
+        description="Determine if the files in folder are being compressed. If True "
+        "all the files are compressed. If 'atomate' only a selection of files related "
+        "to the simulation will be compressed. If False no file is compressed.",
+    )
+    VASP_INHERIT_INCAR: bool = Field(
+        default=False,
+        description="Whether to inherit INCAR settings from previous calculation. "
+        "This might be useful to port Custodian fixes to child jobs but can also be "
+        "dangerous e.g. when switching from GGA to meta-GGA or relax to static jobs."
+        "Can be overridden on a per-job basis via the inherit_incar keyword of "
+        "VaspInputGenerator.",
+    )
+
+    LOBSTER_CMD: str = Field(
+        default="lobster", description="Command to run standard version of VASP."
+    )
+
+    LOBSTER_CUSTODIAN_MAX_ERRORS: int = Field(
+        5, description="Maximum number of errors to correct before custodian gives up"
+    )
+
+    LOBSTER_ZIP_FILES: Union[bool, Literal["atomate"]] = Field(
+        "atomate",
+        description="Determine if the files in folder are being compressed. If True "
+        "all the files are compressed. If 'atomate' only a selection of files related "
+        "to the simulation will be compressed. If False no file is compressed.",
+    )
+
+    CP2K_CMD: str = Field(
+        "cp2k.psmp", description="Command to run the MPI version of cp2k"
+    )
+    CP2K_RUN_BADER: bool = Field(
+        default=False,
+        description="Whether to run the Bader program when parsing CP2K calculations."
+        "Requires the bader executable to be on the path.",
+    )
+    CP2K_INPUT_UPDATES: dict = Field(
+        default_factory=dict, description="Updates to apply to cp2k input files."
+    )
+    CP2K_RELAX_MAX_FORCE: float = Field(
+        0.25,
+        description="Maximum force allowed on each atom for successful structure "
+        "optimization",
+    )
+    CP2K_VOLUME_CHANGE_WARNING_TOL: float = Field(
+        0.2,
+        description="Maximum volume change allowed in CP2K relaxations before the "
+        "calculation is tagged with a warning",
+    )
+    CP2K_HANDLE_UNSUCCESSFUL: Union[str, bool] = Field(
+        "error",
+        description="Three-way toggle on what to do if the job looks OK but is actually"
+        " unconverged (either electronic or ionic). - True: mark job as COMPLETED, but "
+        "stop children. - False: do nothing, continue with workflow as normal. 'error':"
+        " throw an error",
+    )
+    CP2K_CUSTODIAN_MAX_ERRORS: int = Field(
+        5, description="Maximum number of errors to correct before custodian gives up"
+    )
+    CP2K_STORE_VOLUMETRIC_DATA: Optional[tuple[str]] = Field(
+        None, description="Store data from these files in database if present"
+    )
+    CP2K_STORE_ADDITIONAL_JSON: bool = Field(
+        default=True,
+        description="Ingest any additional JSON data present into database when "
+        "parsing CP2K directories useful for storing duplicate of FW.json",
+    )
+
+    CP2K_ZIP_FILES: Union[bool, Literal["atomate"]] = Field(
+        default=True,
+        description="Determine if the files in folder are being compressed. If True "
+        "all the files are compressed. If 'atomate' only a selection of files related "
+        "to the simulation will be compressed. If False no file is compressed.",
     )
 
     # Elastic constant settings
@@ -88,17 +181,15 @@ class Atomate2Settings(BaseSettings):
     )
 
     # AMSET settings
-    AMSET_SETTINGS_UPDATE: dict = Field(
+    AMSET_SETTINGS_UPDATE: Optional[dict] = Field(
         None, description="Additional settings applied to AMSET settings file."
     )
 
-    class Config:
-        """Pydantic config settings."""
+    model_config = SettingsConfigDict(env_prefix="atomate2_")
 
-        env_prefix = "atomate2_"
-
-    @root_validator(pre=True)
-    def load_default_settings(cls, values):
+    @model_validator(mode="before")
+    @classmethod
+    def load_default_settings(cls, values) -> dict:
         """
         Load settings from file or environment variables.
 
@@ -109,11 +200,22 @@ class Atomate2Settings(BaseSettings):
         """
         from monty.serialization import loadfn
 
-        config_file_path: str = values.get("CONFIG_FILE", _DEFAULT_CONFIG_FILE_PATH)
+        config_file_path = values.get("CONFIG_FILE", _DEFAULT_CONFIG_FILE_PATH)
+        config_file_path = Path(config_file_path).expanduser()
 
         new_values = {}
-        if Path(config_file_path).expanduser().exists():
-            new_values.update(loadfn(Path(config_file_path).expanduser()))
+        if config_file_path.exists():
+            if config_file_path.stat().st_size == 0:
+                warnings.warn(
+                    f"Using atomate2 config file at {config_file_path} but it's empty",
+                    stacklevel=2,
+                )
+            else:
+                try:
+                    new_values.update(loadfn(config_file_path))
+                except ValueError:
+                    raise SyntaxError(
+                        f"atomate2 config file at {config_file_path} is unparsable"
+                    ) from None
 
-        new_values.update(values)
-        return new_values
+        return {**new_values, **values}
