@@ -1,9 +1,9 @@
 """Schemas for electron-phonon renormalisation documents."""
 
 import logging
-from typing import Dict, List, Tuple
 
 import numpy as np
+from emmet.core.structure import StructureMetadata
 from pydantic import BaseModel, Field
 from pymatgen.core import Structure
 from pymatgen.electronic_structure.bandstructure import BandStructure
@@ -11,27 +11,25 @@ from pymatgen.electronic_structure.core import Spin
 
 logger = logging.getLogger(__name__)
 
-__all__ = ["RawElectronicData", "ElectronPhononRenormalisationDoc"]
-
 
 class RawElectronicData(BaseModel):
     """Raw data used to fit electron-phonon renormalisation."""
 
-    displacement_uuids: List[str] = Field(
+    displacement_uuids: list[str] = Field(
         None, description="UUIDs of the displacement band structure calculations"
     )
-    displacement_dirs: List[str] = Field(
+    displacement_dirs: list[str] = Field(
         None, description="Directories of the displacement band structure calculations"
     )
-    displacement_structures: List[Structure] = Field(
+    displacement_structures: list[Structure] = Field(
         None, description="The electron-phonon displaced structures at each temperature"
     )
-    displacement_cbms: List[List[float]] = Field(
+    displacement_cbms: list[list[float]] = Field(
         None,
         description="Conduction band minima of the displaced structures, as an"
         "array with the shape (ntemps, ncbms)",
     )
-    displacement_vbms: List[List[float]] = Field(
+    displacement_vbms: list[list[float]] = Field(
         None,
         description="Valence band maxima of the displaced structures, as an"
         "array with the shape (ntemps, nvbms)",
@@ -49,12 +47,12 @@ class RawElectronicData(BaseModel):
     bulk_vbm: float = Field(
         None, description="Valence band maximum of the bulk (supercell) structure"
     )
-    bulk_vbm_band_indices: Dict[str, List[int]] = Field(
+    bulk_vbm_band_indices: dict[str, list[int]] = Field(
         None,
         description="Indices of bands that are degenerate at the valence band "
         "maximum (zero indexed) in the bulk (supercell) structure",
     )
-    bulk_cbm_band_indices: Dict[str, List[int]] = Field(
+    bulk_cbm_band_indices: dict[str, list[int]] = Field(
         None,
         description="Indices of bands that are degenerate at the conduction band "
         "minimum (zero indexed) in the bulk (supercell) structure",
@@ -71,7 +69,7 @@ class RawElectronicData(BaseModel):
     )
 
 
-class ElectronPhononRenormalisationDoc(BaseModel):
+class ElectronPhononRenormalisationDoc(StructureMetadata):
     """Electron-phonon band gap renormalisation document."""
 
     structure: Structure = Field(
@@ -79,23 +77,19 @@ class ElectronPhononRenormalisationDoc(BaseModel):
         description="The primitive structure for which the electron-phonon was"
         " calculated",
     )
-    formula_pretty: str = Field(
-        None,
-        description="Cleaned representation of the formula",
-    )
-    temperatures: List[float] = Field(
+    temperatures: list[float] = Field(
         None, description="Temperatures at which electron-phonon coupling was obtained"
     )
-    band_gaps: List[float] = Field(
+    band_gaps: list[float] = Field(
         None, description="Temperature renormalised band gaps"
     )
-    vbms: List[float] = Field(
+    vbms: list[float] = Field(
         None, description="Temperature renormalised valence band maxima"
     )
-    cbms: List[float] = Field(
+    cbms: list[float] = Field(
         None, description="Temperature renormalised conduction band minima"
     )
-    delta_band_gaps: List[float] = Field(
+    delta_band_gaps: list[float] = Field(
         None, description="Change in band gap relative to the bulk structure"
     )
     bulk_band_gap: float = Field(
@@ -110,11 +104,11 @@ class ElectronPhononRenormalisationDoc(BaseModel):
     @classmethod
     def from_band_structures(
         cls,
-        temperatures: List[float],
-        displacement_band_structures: List[BandStructure],
-        displacement_structures: List[Structure],
-        displacement_uuids: List[str],
-        displacement_dirs: List[str],
+        temperatures: list[float],
+        displacement_band_structures: list[BandStructure],
+        displacement_structures: list[Structure],
+        displacement_uuids: list[str],
+        displacement_dirs: list[str],
         bulk_band_structure: BandStructure,
         bulk_structure: Structure,
         bulk_uuid: str,
@@ -122,7 +116,7 @@ class ElectronPhononRenormalisationDoc(BaseModel):
         elph_uuid: str,
         elph_dir: str,
         original_structure: Structure,
-    ):
+    ) -> "ElectronPhononRenormalisationDoc":
         """
         Calculate an electron-phonon renormalisation document from band structures.
 
@@ -214,9 +208,9 @@ class ElectronPhononRenormalisationDoc(BaseModel):
         band_gaps = cbms - vbms
         delta_band_gaps = band_gaps - bulk_band_gap
 
-        return cls(
+        return cls.from_structure(
             structure=original_structure,
-            formula_pretty=original_structure.composition.reduced_formula,
+            meta_structure=original_structure,
             temperatures=temperatures,
             band_gaps=band_gaps.tolist(),
             vbms=vbms.tolist(),
@@ -243,10 +237,10 @@ class ElectronPhononRenormalisationDoc(BaseModel):
 
 
 def _get_displacement_band_edges(
-    band_structures: List[BandStructure],
-    band_indices: Dict[Spin, List[int]],
+    band_structures: list[BandStructure],
+    band_indices: dict[Spin, list[int]],
     cbm: bool = True,
-):
+) -> np.ndarray:
     """Extract band edge energies based on a band structure and band indices."""
     band_edges = []
     for band_structure in band_structures:
@@ -265,7 +259,7 @@ def _get_displacement_band_edges(
 def _get_band_edge_indices(
     band_structure: BandStructure,
     tol: float = 0.005,
-) -> Tuple[Dict[Spin, List[int]], Dict[Spin, List[int]]]:
+) -> tuple[dict[Spin, list[int]], dict[Spin, list[int]]]:
     """
     Get indices of degenerate band edge states, within a tolerance.
 
