@@ -16,8 +16,6 @@ import paramiko
 from monty.io import zopen
 from paramiko import SFTPClient, SSHClient
 
-__all__ = ["FileClient", "auto_fileclient"]
-
 
 class FileClient:
     """
@@ -43,13 +41,13 @@ class FileClient:
         self,
         key_filename: str | Path = "~/.ssh/id_rsa",
         config_filename: str | Path = "~/.ssh/config",
-    ):
+    ) -> None:
         self.key_filename = key_filename
         self.config_filename = config_filename
 
         self.connections: dict[str, dict[str, Any]] = {}
 
-    def connect(self, host: str):
+    def connect(self, host: str) -> None:
         """
         Connect to a remote host.
 
@@ -135,13 +133,12 @@ class FileClient:
         """
         if host is None:
             return Path(path).exists()
-        else:
-            path = str(self.abspath(path, host=host))
-            try:
-                self.get_sftp(host).stat(path)
-                return True
-            except FileNotFoundError:
-                return False
+        path = str(self.abspath(path, host=host))
+        try:
+            self.get_sftp(host).stat(path)
+            return True
+        except FileNotFoundError:
+            return False
 
     def is_file(self, path: str | Path, host: str | None = None) -> bool:
         """
@@ -161,12 +158,11 @@ class FileClient:
         """
         if host is None:
             return Path(path).is_file()
-        else:
-            path = str(self.abspath(path, host=host))
-            try:
-                return stat.S_ISREG(self.get_sftp(host).lstat(path).st_mode)
-            except FileNotFoundError:
-                return False
+        path = str(self.abspath(path, host=host))
+        try:
+            return stat.S_ISREG(self.get_sftp(host).lstat(path).st_mode)
+        except FileNotFoundError:
+            return False
 
     def is_dir(self, path: str | Path, host: str | None = None) -> bool:
         """
@@ -186,12 +182,11 @@ class FileClient:
         """
         if host is None:
             return Path(path).is_dir()
-        else:
-            path = str(self.abspath(path, host=host))
-            try:
-                return stat.S_ISDIR(self.get_sftp(host).lstat(path).st_mode)
-            except FileNotFoundError:
-                return False
+        path = str(self.abspath(path, host=host))
+        try:
+            return stat.S_ISDIR(self.get_sftp(host).lstat(path).st_mode)
+        except FileNotFoundError:
+            return False
 
     def listdir(self, path: str | Path, host: str | None = None) -> list[Path]:
         """
@@ -213,9 +208,8 @@ class FileClient:
             path = self.abspath(path, host=host)
             return [p.relative_to(path) for p in Path(path).iterdir()]
 
-        else:
-            path = str(self.abspath(path, host=host))
-            return [Path(p) for p in self.get_sftp(host).listdir(path)]
+        path = str(self.abspath(path, host=host))
+        return [Path(p) for p in self.get_sftp(host).listdir(path)]
 
     def copy(
         self,
@@ -223,7 +217,7 @@ class FileClient:
         dest_filename: str | Path,
         src_host: str | None = None,
         dest_host: str | None = None,
-    ):
+    ) -> None:
         """
         Copy a file from source to destination.
 
@@ -255,7 +249,7 @@ class FileClient:
             ssh = self.get_ssh(src_host)
             _, _, stderr = ssh.exec_command(f"cp {src_filename} {dest_filename}")
             if len(stderr.readlines()) > 0:
-                warnings.warn(f"Copy command gave error: {stderr}")
+                warnings.warn(f"Copy command gave error: {stderr}", stacklevel=2)
         else:
             # copying between two remote hosts; this is a pain and it is unlikely anyone
             # will want to do it.
@@ -263,7 +257,7 @@ class FileClient:
                 "Copying between two different remote hosts is not supported."
             )
 
-    def remove(self, path: str | Path, host: str | None = None):
+    def remove(self, path: str | Path, host: str | None = None) -> None:
         """
         Remove a file (does not work on directories).
 
@@ -285,7 +279,7 @@ class FileClient:
         old_path: str | Path,
         new_path: str | Path,
         host: str | None = None,
-    ):
+    ) -> None:
         """
         Rename (move) a file.
 
@@ -323,10 +317,9 @@ class FileClient:
         """
         if host is None:
             return Path(path).absolute()
-        else:
-            ssh = self.get_ssh(host)
-            _, stdout, _ = ssh.exec_command(f"readlink -f {path}")
-            return Path([o.split("\n")[0] for o in stdout][0])
+        ssh = self.get_ssh(host)
+        _, stdout, _ = ssh.exec_command(f"readlink -f {path}")
+        return Path(next(o.split("\n")[0] for o in stdout))
 
     def glob(self, path: str | Path, host: str | None = None) -> list[Path]:
         """
@@ -358,8 +351,8 @@ class FileClient:
         path: str | Path,
         host: str | None = None,
         compresslevel: int = 6,
-        force: bool = False,
-    ):
+        force: bool | str = False,
+    ) -> None:
         """
         Gzip a file.
 
@@ -372,21 +365,39 @@ class FileClient:
         compresslevel : bool
             Level of compression, 1-9. 9 is default for GzipFile, 6 is default for gzip.
         force : bool
-            Overwrite gzipped file if it already exists.
+            How to handle writing a gzipped file if it already exists. Accepts
+            either a string or bool:
+
+            - `"force"` or `True`: Overwrite gzipped file if it already exists.
+            - `"raise"` or `False`: Raise an error if file already exists.
+            - `"skip"` Skip file if it already exists.
         """
         path = self.abspath(path, host=host)
         path_gz = path.parent / f"{path.name}.gz"
 
         if str(path).lower().endswith("gz"):
-            warnings.warn(f"{path} is already gzipped, skipping...")
-            return None
+            warnings.warn(f"{path} is already gzipped, skipping...", stacklevel=1)
+            return
 
         if self.is_dir(path, host=host):
-            warnings.warn(f"{path} is a directory, skipping...")
-            return None
+            warnings.warn(f"{path} is a directory, skipping...", stacklevel=1)
+            return
 
-        if self.exists(path_gz, host=host) and not force:
-            raise FileExistsError(f"{path_gz} file already exists.")
+        if self.exists(path_gz, host=host):
+            if force is False or force == "raise":
+                raise FileExistsError(f"{path_gz} file already exists")
+            if force is True or force == "force":
+                pass
+            elif force == "skip":
+                warnings.warn(
+                    f"{path_gz} file already exists, skipping...", stacklevel=2
+                )
+                return
+            else:
+                raise ValueError(
+                    f"Invalid value for force: {force} "
+                    "(must be True, False, 'raise', 'force', or 'skip'))"
+                )
 
         if host is None:
             with open(path, "rb") as f_in, GzipFile(
@@ -397,14 +408,14 @@ class FileClient:
             path.unlink()
         else:
             ssh = self.get_ssh(host)
-            _, stdout, _ = ssh.exec_command(f"gzip -f {str(path)}")
+            _, stdout, _ = ssh.exec_command(f"gzip -f {path!s}")
 
     def gunzip(
         self,
         path: str | Path,
         host: str | None = None,
-        force: bool = False,
-    ):
+        force: bool | str = False,
+    ) -> None:
         """
         Ungzip a file.
 
@@ -415,17 +426,35 @@ class FileClient:
         host : str or None
             A remote file system host on which to perform file operations.
         force : bool
-            Overwrite non-gzipped file if it already exists.
+            How to handle writing a non-gzipped file if it already exists. Accepts
+            either a string or bool:
+
+            - `"force"` or `True`: Overwrite non-gzipped file if it already exists.
+            - `"raise"` or `False`: Raise an error if file already exists.
+            - `"skip"` Skip file if it already exists.
         """
         path = self.abspath(path, host=host)
         path_nongz = path.with_suffix("")
 
         if not str(path).lower().endswith("gz"):
-            warnings.warn(f"{path} is not gzipped, skipping...")
-            return None
+            warnings.warn(f"{path} is not gzipped, skipping...", stacklevel=2)
+            return
 
-        if self.exists(path_nongz, host=host) and not force:
-            raise FileExistsError(f"{path_nongz} file already exists")
+        if self.exists(path_nongz, host=host):
+            if force is False or force == "raise":
+                raise FileExistsError(f"{path_nongz} file already exists")
+            if force is True or force == "force":
+                pass
+            elif force == "skip":
+                warnings.warn(
+                    f"{path_nongz} file already exists, skipping...", stacklevel=2
+                )
+                return
+            else:
+                raise ValueError(
+                    f"Invalid value for force: {force} "
+                    "(must be True, False, 'raise', 'force', or 'skip'))"
+                )
 
         if host is None:
             with open(path_nongz, "wb") as f_out, zopen(path, "rb") as f_in:
@@ -433,9 +462,9 @@ class FileClient:
             path.unlink()
         else:
             ssh = self.get_ssh(host)
-            _, stdout, _ = ssh.exec_command(f"gunzip -f {str(path)}")
+            _, stdout, _ = ssh.exec_command(f"gunzip -f {path!s}")
 
-    def close(self):
+    def close(self) -> None:
         """Close all connections."""
         for connection in self.connections.values():
             connection["ssh"].close()
@@ -446,7 +475,7 @@ class FileClient:
         """Support for "with" context."""
         return self
 
-    def __exit__(self, exc_type, exc_val, exc_tb):
+    def __exit__(self, exc_type, exc_val, exc_tb) -> None:
         """Support for "with" context."""
         self.close()
 
@@ -488,7 +517,7 @@ def get_ssh_connection(
         # try reading ssh config file
         ssh_config = paramiko.SSHConfig().from_path(str(config_filename))
 
-        host_config = ssh_config.lookup(hostname)
+        host_config = ssh_config.lookup(hostname)  # type: ignore[attr-defined]
         for k in ("hostname", "user", "port"):
             if k in host_config:
                 config[k.replace("user", "username")] = host_config[k]
@@ -521,7 +550,7 @@ def auto_fileclient(method: Callable | None = None):
     def decorator(func):
         @wraps(func)
         def gen_fileclient(*args, **kwargs):
-            file_client = kwargs.get("file_client", None)
+            file_client = kwargs.get("file_client")
             if file_client is None:
                 with FileClient() as file_client:
                     kwargs["file_client"] = file_client
