@@ -2,11 +2,10 @@
 
 from __future__ import annotations
 
+import contextlib
 import os
 import socket
 from pathlib import Path
-
-__all__ = ["get_uri", "strip_hostname"]
 
 
 def get_uri(dir_name: str | Path) -> str:
@@ -27,10 +26,8 @@ def get_uri(dir_name: str | Path) -> str:
     """
     fullpath = Path(dir_name).absolute()
     hostname = socket.gethostname()
-    try:
+    with contextlib.suppress(socket.gaierror, socket.herror):
         hostname = socket.gethostbyaddr(hostname)[0]
-    except (socket.gaierror, socket.herror):
-        pass
     return f"{hostname}:{fullpath}"
 
 
@@ -57,7 +54,9 @@ def strip_hostname(uri_path: str | Path) -> str:
     return dir_name
 
 
-def find_recent_logfile(dir_name: Path | str, logfile_extensions: str | list[str]):
+def find_recent_logfile(
+    dir_name: Path | str, logfile_extensions: str | list[str]
+) -> str:
     """
     Find the most recent logfile in a given directory.
 
@@ -76,7 +75,7 @@ def find_recent_logfile(dir_name: Path | str, logfile_extensions: str | list[str
     """
     mod_time = 0.0
     logfile = None
-    if type(logfile_extensions) == str:
+    if isinstance(logfile_extensions, str):
         logfile_extensions = [logfile_extensions]
     for f in os.listdir(dir_name):
         f_path = os.path.join(dir_name, f)
