@@ -3,28 +3,20 @@
 from __future__ import annotations
 
 from dataclasses import dataclass, field
-from pathlib import Path
-
-from jobflow import Flow, Maker, OnMissing
-from pymatgen.core.structure import Structure
+from typing import TYPE_CHECKING
 
 from atomate2 import SETTINGS
-from atomate2.common.schemas.math import Matrix3D
+from atomate2.common.flows.elastic import BaseElasticMaker
 from atomate2.vasp.flows.core import DoubleRelaxMaker
-from atomate2.vasp.jobs.base import BaseVaspMaker
 from atomate2.vasp.jobs.core import TightRelaxMaker
-from atomate2.vasp.jobs.elastic import (
-    ElasticRelaxMaker,
-    fit_elastic_tensor,
-    generate_elastic_deformations,
-    run_elastic_deformations,
-)
+from atomate2.vasp.jobs.elastic import ElasticRelaxMaker
 
-__all__ = ["ElasticMaker"]
+if TYPE_CHECKING:
+    from atomate2.vasp.jobs.base import BaseVaspMaker
 
 
 @dataclass
-class ElasticMaker(Maker):
+class ElasticMaker(BaseElasticMaker):
     """
     Maker to calculate elastic constants.
 
@@ -61,6 +53,8 @@ class ElasticMaker(Maker):
         Keyword arguments passed to :obj:`generate_elastic_deformations`.
     fit_elastic_tensor_kwargs : dict
         Keyword arguments passed to :obj:`fit_elastic_tensor`.
+    task_document_kwargs : dict
+        Additional keyword args passed to :obj:`.ElasticDocument.from_stresses()`.
     """
 
     name: str = "elastic"
@@ -73,66 +67,16 @@ class ElasticMaker(Maker):
     elastic_relax_maker: BaseVaspMaker = field(default_factory=ElasticRelaxMaker)
     generate_elastic_deformations_kwargs: dict = field(default_factory=dict)
     fit_elastic_tensor_kwargs: dict = field(default_factory=dict)
+    task_document_kwargs: dict = field(default_factory=dict)
 
-    def make(
-        self,
-        structure: Structure,
-        prev_vasp_dir: str | Path | None = None,
-        equilibrium_stress: Matrix3D = None,
-    ):
+    @property
+    def prev_calc_dir_argname(self):
+        """Name of argument informing static maker of previous calculation directory.
+
+        As this differs between different DFT codes (e.g., VASP, CP2K), it
+        has been left as a property to be implemented by the inheriting class.
+
+        Note: this is only applicable if a relax_maker is specified; i.e., two
+        calculations are performed for each ordering (relax -> static)
         """
-        Make flow to calculate the elastic constant.
-
-        Parameters
-        ----------
-        structure : .Structure
-            A pymatgen structure.
-        prev_vasp_dir : str or Path or None
-            A previous vasp calculation directory to use for copying outputs.
-        equilibrium_stress : tuple of tuple of float
-            The equilibrium stress of the (relaxed) structure, if known.
-        """
-        jobs = []
-
-        if self.bulk_relax_maker is not None:
-            # optionally relax the structure
-            bulk = self.bulk_relax_maker.make(structure, prev_vasp_dir=prev_vasp_dir)
-            jobs.append(bulk)
-            structure = bulk.output.structure
-            prev_vasp_dir = bulk.output.dir_name
-            if equilibrium_stress is None:
-                equilibrium_stress = bulk.output.output.stress
-
-        deformations = generate_elastic_deformations(
-            structure,
-            order=self.order,
-            sym_reduce=self.sym_reduce,
-            symprec=self.symprec,
-            **self.generate_elastic_deformations_kwargs,
-        )
-        vasp_deformation_calcs = run_elastic_deformations(
-            structure,
-            deformations.output,
-            prev_vasp_dir=prev_vasp_dir,
-            elastic_relax_maker=self.elastic_relax_maker,
-        )
-        fit_tensor = fit_elastic_tensor(
-            structure,
-            vasp_deformation_calcs.output,
-            equilibrium_stress=equilibrium_stress,
-            order=self.order,
-            symprec=self.symprec if self.sym_reduce else None,
-            **self.fit_elastic_tensor_kwargs,
-        )
-
-        # allow some of the deformations to fail
-        fit_tensor.config.on_missing_references = OnMissing.NONE
-
-        jobs += [deformations, vasp_deformation_calcs, fit_tensor]
-
-        flow = Flow(
-            jobs=jobs,
-            output=fit_tensor.output,
-            name=self.name,
-        )
-        return flow
+        return "prev_dir"
