@@ -1,19 +1,13 @@
-"""Flows for calculating elastic constants."""
+"""Flows for calculating phonons."""
 
 from __future__ import annotations
 
 from dataclasses import dataclass, field
-from pathlib import Path
+from typing import TYPE_CHECKING
 
 from jobflow import Flow, Maker
-from pymatgen.core.structure import Structure
 
-from atomate2.common.jobs import structure_to_conventional, structure_to_primitive
-from atomate2.common.schemas.math import Matrix3D
-from atomate2.vasp.flows.core import DoubleRelaxMaker
-from atomate2.vasp.jobs.base import BaseVaspMaker
-from atomate2.vasp.jobs.core import DielectricMaker, StaticMaker, TightRelaxMaker
-from atomate2.vasp.jobs.phonons import (
+from atomate2.common.jobs.phonons import (
     PhononDisplacementMaker,
     generate_frequencies_eigenvectors,
     generate_phonon_displacements,
@@ -21,8 +15,18 @@ from atomate2.vasp.jobs.phonons import (
     get_total_energy_per_cell,
     run_phonon_displacements,
 )
+from atomate2.common.jobs.utils import structure_to_conventional, structure_to_primitive
+from atomate2.vasp.flows.core import DoubleRelaxMaker
+from atomate2.vasp.jobs.core import DielectricMaker, StaticMaker, TightRelaxMaker
+from atomate2.vasp.sets.core import StaticSetGenerator
 
-__all__ = ["PhononMaker"]
+if TYPE_CHECKING:
+    from pathlib import Path
+
+    from emmet.core.math import Matrix3D
+    from pymatgen.core.structure import Structure
+
+    from atomate2.vasp.jobs.base import BaseVaspMaker
 
 
 @dataclass
@@ -36,7 +40,7 @@ class PhononMaker(Maker):
     forces are computed for these structures. With the help of phonopy, these
     forces are then converted into a dynamical matrix. To correct for polarization
     effects, a correction of the dynamical matrix based on BORN charges can
-    be performed.     Finally, phonon densities of states, phonon band structures
+    be performed. Finally, phonon densities of states, phonon band structures
     and thermodynamic properties are computed.
 
     .. Note::
@@ -49,25 +53,25 @@ class PhononMaker(Maker):
 
     Parameters
     ----------
-    name : str
+    name : str = "phonon"
         Name of the flows produced by this maker.
-    sym_reduce : bool
+    sym_reduce : bool = True
         Whether to reduce the number of deformations using symmetry.
-    symprec : float
+    symprec : float = 1e-4
         Symmetry precision to use in the
         reduction of symmetry to find the primitive/conventional cell
         (use_primitive_standard_structure, use_conventional_standard_structure)
         and to handle all symmetry-related tasks in phonopy
-    displacement: float
+    displacement: float = 0.01
         displacement distance for phonons
-    min_length: float
+    min_length: float = 20.0
         min length of the supercell that will be built
-    prefer_90_degrees: bool
+    prefer_90_degrees: bool = True
         if set to True, supercell algorithm will first try to find a supercell
         with 3 90 degree angles
-    get_supercell_size_kwargs: dict
+    get_supercell_size_kwargs: dict = {}
         kwargs that will be passed to get_supercell_size to determine supercell size
-    use_symmetrized_structure: str
+    use_symmetrized_structure: str or None = None
         allowed strings: "primitive", "conventional", None
 
         - "primitive" will enforce to start the phonon computation
@@ -102,7 +106,7 @@ class PhononMaker(Maker):
         Keyword arguments passed to :obj:`generate_frequencies_eigenvectors`.
     create_thermal_displacements: bool
         Bool that determines if thermal_displacement_matrices are computed
-    kpath_scheme: str
+    kpath_scheme: str = "seekpath"
         scheme to generate kpoints. Please be aware that
         you can only use seekpath with any kind of cell
         Otherwise, please use the standard primitive structure
@@ -112,8 +116,8 @@ class PhononMaker(Maker):
         seekpath can be used with any kind of unit cell as
         it relies on phonopy to handle the relationship
         to the primitive cell and not pymatgen
-    code: str
-        determines the dft code. currently only vasp is implemented.
+    code: str = "vasp"
+        determines the DFT code. currently only vasp is implemented.
         This keyword might enable the implementation of other codes
         in the future
     store_force_constants: bool
@@ -131,7 +135,11 @@ class PhononMaker(Maker):
     bulk_relax_maker: BaseVaspMaker | None = field(
         default_factory=lambda: DoubleRelaxMaker.from_relax_maker(TightRelaxMaker())
     )
-    static_energy_maker: BaseVaspMaker | None = field(default_factory=StaticMaker)
+    static_energy_maker: BaseVaspMaker | None = field(
+        default_factory=lambda: StaticMaker(
+            input_set_generator=StaticSetGenerator(auto_ispin=True)
+        )
+    )
     born_maker: BaseVaspMaker | None = field(default_factory=DielectricMaker)
     phonon_displacement_maker: BaseVaspMaker = field(
         default_factory=PhononDisplacementMaker
@@ -145,46 +153,41 @@ class PhononMaker(Maker):
     def make(
         self,
         structure: Structure,
-        prev_vasp_dir: str | Path | None = None,
+        prev_dir: str | Path | None = None,
         born: list[Matrix3D] | None = None,
         epsilon_static: Matrix3D | None = None,
         total_dft_energy_per_formula_unit: float | None = None,
         supercell_matrix: Matrix3D | None = None,
-    ):
+    ) -> Flow:
         """
         Make flow to calculate the phonon properties.
 
         Parameters
         ----------
         structure : .Structure
-            A pymatgen structure. Please start with a structure
-            that is nearly fully optimized as the internal optimizers
-            have very strict settings!
-        prev_vasp_dir : str or Path or None
+            A pymatgen structure. Please start with a structure that is nearly fully
+            optimized as the internal optimizers have very strict settings!
+        prev_dir : str or Path or None
             A previous vasp calculation directory to use for copying outputs.
         born: Matrix3D
-            Instead of recomputing born charges and epsilon,
-            these values can also be provided manually.
-            if born, epsilon_static are provided, the born
-            run will be skipped
-            it can be provided in the VASP convention with information for
-            every atom in unit cell. Please be careful when converting
-            structures within in this workflow as this could lead to errors
+            Instead of recomputing born charges and epsilon, these values can also be
+            provided manually. If born and epsilon_static are provided, the born run
+            will be skipped it can be provided in the VASP convention with information
+            for every atom in unit cell. Please be careful when converting structures
+            within in this workflow as this could lead to errors
         epsilon_static: Matrix3D
-            The high-frequency dielectric constant
-            Instead of recomputing born charges and epsilon,
-            these values can also be provided.
-            if born, epsilon_static are provided, the born
-            run will be skipped
+            The high-frequency dielectric constant to use instead of recomputing born
+            charges and epsilon. If born, epsilon_static are provided, the born run
+            will be skipped
         total_dft_energy_per_formula_unit: float
-            It has to be given per formula unit (as a result in corresponding Doc)
-            Instead of recomputing the energy of the bulk structure every time,
-            this value can also be provided in eV. If it is provided,
-            the static run will be skipped. This energy is the typical
-            output dft energy of the dft workflow. No conversion needed.
+            It has to be given per formula unit (as a result in corresponding Doc).
+            Instead of recomputing the energy of the bulk structure every time, this
+            value can also be provided in eV. If it is provided, the static run will be
+            skipped. This energy is the typical output DFT energy of the DFT workflow.
+            No conversion needed.
         supercell_matrix: list
-            instead of min_length, also a supercell_matrix can
-            be given, e.g. [[1.0,0.0,0.0],[0.0,1.0,0.0],[0.0,0.0,1.0]
+            Instead of min_length, also a supercell_matrix can be given, e.g.
+            [[1.0,0.0,0.0],[0.0,1.0,0.0],[0.0,0.0,1.0]
         """
         if self.use_symmetrized_structure not in [None, "primitive", "conventional"]:
             raise ValueError(
@@ -192,28 +195,25 @@ class PhononMaker(Maker):
             )
 
         if (
-            not self.use_symmetrized_structure == "primitive"
+            self.use_symmetrized_structure != "primitive"
             and self.kpath_scheme != "seekpath"
         ):
             raise ValueError(
-                "You can only use other kpath schemes with the primitive standard structure"
+                "You can only use other kpath schemes with the primitive standard "
+                "structure"
             )
 
-        if self.kpath_scheme not in [
-            "seekpath",
-            "hinuma",
-            "setyawan_curtarolo",
-            "latimer_munro",
-        ]:
+        if (
+            self.kpath_scheme
+            not in "seekpath hinuma setyawan_curtarolo latimer_munro".split()
+        ):
             raise ValueError("kpath scheme is not implemented")
 
         jobs = []
 
-        # TODO: should this be after or before structural
-        # optimization as the optimization could change
-        # the symmetry
-        # we could add a tutorial and point out that the structure
-        # should be nearly optimized before the phonon workflow
+        # TODO: should this be after or before structural optimization as the
+        #  optimization could change the symmetry we could add a tutorial and point out
+        #  that the structure should be nearly optimized before the phonon workflow
         if self.use_symmetrized_structure == "primitive":
             # These structures are compatible with many
             # of the kpath algorithms that are used for Materials Project
@@ -221,27 +221,25 @@ class PhononMaker(Maker):
             jobs.append(prim_job)
             structure = prim_job.output
         elif self.use_symmetrized_structure == "conventional":
-            # it could be beneficial to use conventional
-            # standard structures to arrive faster at supercells with right
-            # angles
+            # it could be beneficial to use conventional standard structures to arrive
+            # faster at supercells with right angles
             conv_job = structure_to_conventional(structure, self.symprec)
             jobs.append(conv_job)
             structure = conv_job.output
 
+        optimization_run_job_dir = None
+        optimization_run_uuid = None
         if self.bulk_relax_maker is not None:
             # optionally relax the structure
-            bulk = self.bulk_relax_maker.make(structure, prev_vasp_dir=prev_vasp_dir)
+            bulk = self.bulk_relax_maker.make(structure, prev_dir=prev_dir)
             jobs.append(bulk)
             structure = bulk.output.structure
+            prev_dir = bulk.output.dir_name
             optimization_run_job_dir = bulk.output.dir_name
             optimization_run_uuid = bulk.output.uuid
-        else:
-            optimization_run_job_dir = None
-            optimization_run_uuid = None
 
-        # if supercell_matrix is None, supercell size will be determined
-        # after relax maker to ensure that cell lengths are really larger
-        # than threshold
+        # if supercell_matrix is None, supercell size will be determined after relax
+        # maker to ensure that cell lengths are really larger than threshold
         if supercell_matrix is None:
             supercell_job = get_supercell_size(
                 structure,
@@ -251,6 +249,29 @@ class PhononMaker(Maker):
             )
             jobs.append(supercell_job)
             supercell_matrix = supercell_job.output
+
+        # Computation of static energy
+        total_dft_energy = None
+        static_run_job_dir = None
+        static_run_uuid = None
+        if (self.static_energy_maker is not None) and (
+            total_dft_energy_per_formula_unit is None
+        ):
+            static_job = self.static_energy_maker.make(
+                structure=structure, prev_dir=prev_dir
+            )
+            jobs.append(static_job)
+            total_dft_energy = static_job.output.output.energy
+            static_run_job_dir = static_job.output.dir_name
+            static_run_uuid = static_job.output.uuid
+            prev_dir = static_job.output.dir_name
+        elif total_dft_energy_per_formula_unit is not None:
+            # to make sure that one can reuse results from Doc
+            compute_total_energy_job = get_total_energy_per_cell(
+                total_dft_energy_per_formula_unit, structure
+            )
+            jobs.append(compute_total_energy_job)
+            total_dft_energy = compute_total_energy_job.output
 
         # get a phonon object from phonopy
         displacements = generate_phonon_displacements(
@@ -271,34 +292,15 @@ class PhononMaker(Maker):
             structure=structure,
             supercell_matrix=supercell_matrix,
             phonon_maker=self.phonon_displacement_maker,
+            prev_dir=prev_dir,
         )
         jobs.append(vasp_displacement_calcs)
 
-        # Computation of static energy
-        if (self.static_energy_maker is not None) and (
-            total_dft_energy_per_formula_unit is None
-        ):
-            static_job = self.static_energy_maker.make(structure=structure)
-            jobs.append(static_job)
-            total_dft_energy = static_job.output.output.energy
-            static_run_job_dir = static_job.output.dir_name
-            static_run_uuid = static_job.output.uuid
-        else:
-            if total_dft_energy_per_formula_unit is not None:
-                # to make sure that one can reuse results from Doc
-                compute_total_energy_job = get_total_energy_per_cell(
-                    total_dft_energy_per_formula_unit, structure
-                )
-                jobs.append(compute_total_energy_job)
-                total_dft_energy = compute_total_energy_job.output
-            else:
-                total_dft_energy = None
-            static_run_job_dir = None
-            static_run_uuid = None
-
         # Computation of BORN charges
+        born_run_job_dir = None
+        born_run_uuid = None
         if self.born_maker is not None and (born is None or epsilon_static is None):
-            born_job = self.born_maker.make(structure)
+            born_job = self.born_maker.make(structure, prev_dir=prev_dir)
             jobs.append(born_job)
 
             # I am not happy how we currently access "born" charges
@@ -307,9 +309,6 @@ class PhononMaker(Maker):
             born = born_job.output.calcs_reversed[0].output.outcar["born"]
             born_run_job_dir = born_job.output.dir_name
             born_run_uuid = born_job.output.uuid
-        else:
-            born_run_job_dir = None
-            born_run_uuid = None
 
         phonon_collect = generate_frequencies_eigenvectors(
             supercell_matrix=supercell_matrix,
@@ -334,8 +333,7 @@ class PhononMaker(Maker):
             store_force_constants=self.store_force_constants,
             **self.generate_frequencies_eigenvectors_kwargs,
         )
-
         jobs.append(phonon_collect)
+
         # create a flow including all jobs for a phonon computation
-        flow = Flow(jobs, phonon_collect.output)
-        return flow
+        return Flow(jobs, phonon_collect.output)
