@@ -5,31 +5,20 @@ from __future__ import annotations
 import logging
 from copy import deepcopy
 from dataclasses import dataclass, field
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 import numpy as np
-from pymatgen.core import Structure
 from pymatgen.core.periodic_table import Element
-from pymatgen.io.vasp import Outcar, Vasprun
 
-from atomate2.common.schemas.math import Vector3D
 from atomate2.vasp.sets.base import VaspInputGenerator
 
+if TYPE_CHECKING:
+    from emmet.core.math import Vector3D
+    from pymatgen.core import Structure
+    from pymatgen.io.vasp import Outcar, Vasprun
+
+
 logger = logging.getLogger(__name__)
-
-
-__all__ = [
-    "RelaxSetGenerator",
-    "TightRelaxSetGenerator",
-    "StaticSetGenerator",
-    "NonSCFSetGenerator",
-    "HSERelaxSetGenerator",
-    "HSEStaticSetGenerator",
-    "HSEBSSetGenerator",
-    "HSETightRelaxSetGenerator",
-    "ElectronPhononSetGenerator",
-    "MDSetGenerator",
-]
 
 
 @dataclass
@@ -40,7 +29,7 @@ class RelaxSetGenerator(VaspInputGenerator):
         self,
         structure: Structure,
         prev_incar: dict = None,
-        bandgap: float = 0,
+        bandgap: float = None,
         vasprun: Vasprun = None,
         outcar: Outcar = None,
     ) -> dict:
@@ -76,7 +65,7 @@ class TightRelaxSetGenerator(VaspInputGenerator):
         self,
         structure: Structure,
         prev_incar: dict = None,
-        bandgap: float = 0,
+        bandgap: float = None,
         vasprun: Vasprun = None,
         outcar: Outcar = None,
     ) -> dict:
@@ -138,7 +127,7 @@ class StaticSetGenerator(VaspInputGenerator):
         self,
         structure: Structure,
         prev_incar: dict = None,
-        bandgap: float = 0,
+        bandgap: float = None,
         vasprun: Vasprun = None,
         outcar: Outcar = None,
     ) -> dict:
@@ -163,13 +152,7 @@ class StaticSetGenerator(VaspInputGenerator):
         dict
             A dictionary of updates to apply.
         """
-        updates = {
-            "NSW": 0,
-            "ISMEAR": -5,
-            "LCHARG": True,
-            "LORBIT": 11,
-            "LREAL": False,
-        }
+        updates = {"NSW": 0, "ISMEAR": -5, "LCHARG": True, "LORBIT": 11, "LREAL": False}
         if self.lepsilon:
             # LPEAD=T: numerical evaluation of overlap integral prevents LRF_COMMUTATOR
             # errors and can lead to better expt. agreement but produces slightly
@@ -194,6 +177,9 @@ class NonSCFSetGenerator(VaspInputGenerator):
         Energy difference used to set NEDOS, based on the total energy range.
     reciprocal_density
         Density of k-mesh by reciprocal volume.
+    reciprocal_density_metal
+        Density of k-mesh by reciprocal volume for use when the system is metallic
+        and ``auto_metal_kpoints=True`` (the default).
     line_density
         Line density for line mode band structure.
     optics
@@ -208,12 +194,13 @@ class NonSCFSetGenerator(VaspInputGenerator):
     mode: str = "line"
     dedos: float = 0.02
     reciprocal_density: float = 100
+    reciprocal_density_metal: float = 400
     line_density: float = 20
     optics: bool = False
     nbands_factor: float = 1.2
     auto_ispin: bool = True
 
-    def __post_init__(self):
+    def __post_init__(self) -> None:
         """Ensure mode is set correctly."""
         super().__post_init__()
         self.mode = self.mode.lower()
@@ -256,16 +243,19 @@ class NonSCFSetGenerator(VaspInputGenerator):
         if self.mode == "line":
             return {"line_density": self.line_density}
 
-        elif self.mode == "boltztrap":
+        if self.mode == "boltztrap":
             return {"explicit": True, "reciprocal_density": self.reciprocal_density}
 
-        return {"reciprocal_density": self.reciprocal_density}
+        return {
+            "reciprocal_density": self.reciprocal_density,
+            "reciprocal_density_metal": self.reciprocal_density_metal,
+        }
 
     def get_incar_updates(
         self,
         structure: Structure,
         prev_incar: dict = None,
-        bandgap: float = 0,
+        bandgap: float = None,
         vasprun: Vasprun = None,
         outcar: Outcar = None,
     ) -> dict:
@@ -301,21 +291,23 @@ class NonSCFSetGenerator(VaspInputGenerator):
         }
 
         if vasprun is not None:
-            # set nbands
-            nbands = int(np.ceil(vasprun.parameters["NBANDS"] * self.nbands_factor))
-            updates["NBANDS"] = nbands
+            # set NBANDS
+            n_bands = int(np.ceil(vasprun.parameters["NBANDS"] * self.nbands_factor))
+            updates["NBANDS"] = n_bands
 
         if self.mode == "uniform":
-            # automatic setting of nedos using the energy range and the energy step
-            nedos = _get_nedos(vasprun, self.dedos)
+            # automatic setting of NEDOS using the energy range and the energy step
+            n_edos = _get_nedos(vasprun, self.dedos)
 
             # use tetrahedron method for DOS and optics calculations
-            updates.update({"ISMEAR": -5, "ISYM": 2, "NEDOS": nedos})
+            updates.update({"ISMEAR": -5, "ISYM": 2, "NEDOS": n_edos})
 
         elif self.mode in ("line", "boltztrap"):
             # if line mode or explicit k-points (boltztrap) can't use ISMEAR=-5
             # use small sigma to avoid partial occupancies for small band gap materials
-            updates.update({"ISMEAR": 0, "SIGMA": 0.01})
+            # use a larger sigma if the material is a metal
+            sigma = 0.2 if bandgap == 0 else 0.01
+            updates.update({"ISMEAR": 0, "SIGMA": sigma})
 
         if self.optics:
             # LREAL not supported with LOPTICS = True; automatic NEDOS usually
@@ -337,14 +329,13 @@ class HSERelaxSetGenerator(VaspInputGenerator):
         By default the hybrid input sets use ALGO = Normal which is only efficient for
         VASP 6.0 and higher. See https://www.vasp.at/wiki/index.php/LFOCKACE for more
         details.
-
     """
 
     def get_incar_updates(
         self,
         structure: Structure,
         prev_incar: dict = None,
-        bandgap: float = 0,
+        bandgap: float = None,
         vasprun: Vasprun = None,
         outcar: Outcar = None,
     ) -> dict:
@@ -397,7 +388,7 @@ class HSETightRelaxSetGenerator(VaspInputGenerator):
         self,
         structure: Structure,
         prev_incar: dict = None,
-        bandgap: float = 0,
+        bandgap: float = None,
         vasprun: Vasprun = None,
         outcar: Outcar = None,
     ) -> dict:
@@ -450,14 +441,13 @@ class HSEStaticSetGenerator(VaspInputGenerator):
         By default the hybrid input sets use ALGO = Normal which is only efficient for
         VASP 6.0 and higher. See https://www.vasp.at/wiki/index.php/LFOCKACE for more
         details.
-
     """
 
     def get_incar_updates(
         self,
         structure: Structure,
         prev_incar: dict = None,
-        bandgap: float = 0,
+        bandgap: float = None,
         vasprun: Vasprun = None,
         outcar: Outcar = None,
     ) -> dict:
@@ -560,7 +550,7 @@ class HSEBSSetGenerator(VaspInputGenerator):
     added_kpoints: list[Vector3D] = field(default_factory=list)
     auto_ispin: bool = True
 
-    def __post_init__(self):
+    def __post_init__(self) -> None:
         """Ensure mode is set correctly."""
         super().__post_init__()
 
@@ -626,7 +616,7 @@ class HSEBSSetGenerator(VaspInputGenerator):
         self,
         structure: Structure,
         prev_incar: dict = None,
-        bandgap: float = 0,
+        bandgap: float = None,
         vasprun: Vasprun = None,
         outcar: Outcar = None,
     ) -> dict:
@@ -729,7 +719,7 @@ class ElectronPhononSetGenerator(VaspInputGenerator):
         self,
         structure: Structure,
         prev_incar: dict = None,
-        bandgap: float = 0,
+        bandgap: float = None,
         vasprun: Vasprun = None,
         outcar: Outcar = None,
     ) -> dict:
@@ -840,7 +830,7 @@ class MDSetGenerator(VaspInputGenerator):
         self,
         structure: Structure,
         prev_incar: dict = None,
-        bandgap: float = 0,
+        bandgap: float = None,
         vasprun: Vasprun = None,
         outcar: Outcar = None,
     ) -> dict:
@@ -886,21 +876,18 @@ class MDSetGenerator(VaspInputGenerator):
             }
         )
 
-        if Element("H") in structure.species:
-            if updates["POTIM"] > 0.5:
-                logger.warning(
-                    f"Molecular dynamics time step is {updates['POTIM']}, which is "
-                    "typically too large for a structure containing H. Consider set it "
-                    "to a value of 0.5 or smaller."
-                )
+        if Element("H") in structure.species and updates["POTIM"] > 0.5:
+            logger.warning(
+                f"Molecular dynamics time step is {updates['POTIM']}, which is "
+                "typically too large for a structure containing H. Consider set it "
+                "to a value of 0.5 or smaller."
+            )
 
         return updates
 
     @staticmethod
     def _get_ensemble_defaults(structure: Structure, ensemble: str) -> dict[str, Any]:
-        """
-        Get default params for the ensemble.
-        """
+        """Get default params for the ensemble."""
         defaults = {
             "nve": {"MDALGO": 1, "ISIF": 2, "ANDERSEN_PROB": 0.0},
             "nvt": {"MDALGO": 2, "ISIF": 2, "SMASS": 0},
@@ -915,15 +902,13 @@ class MDSetGenerator(VaspInputGenerator):
         }
 
         try:
-            return defaults[ensemble.lower()]  # type: ignore
-        except KeyError:
-            supported = tuple(defaults.keys())
-            raise ValueError(
-                f"Expect `ensemble` to be one of {supported}; got {ensemble}."
-            )
+            return defaults[ensemble.lower()]  # type: ignore[return-value]
+        except KeyError as err:
+            supported = tuple(defaults)
+            raise ValueError(f"Expect {ensemble=} to be one of {supported}") from err
 
 
-def _get_nedos(vasprun: Vasprun | None, dedos: float):
+def _get_nedos(vasprun: Vasprun | None, dedos: float) -> int:
     """Automatic setting of nedos using the energy range and the energy step."""
     if vasprun is None:
         return 2000
