@@ -3,6 +3,9 @@ import os
 import shutil
 
 import pytest
+from monty.json import MontyDecoder, jsanitize
+
+from atomate2.common.schemas.cclib import TaskDocument
 
 try:
     import cclib
@@ -12,10 +15,6 @@ except ImportError:
 
 @pytest.mark.skipif(cclib is None, reason="requires cclib to be installed")
 def test_cclib_taskdoc(test_dir):
-    from monty.json import MontyDecoder, jsanitize
-
-    from atomate2.common.schemas.cclib import TaskDocument
-
     p = test_dir / "schemas"
 
     # Plain parsing of task doc. We do not check all cclib entries
@@ -23,31 +22,28 @@ def test_cclib_taskdoc(test_dir):
     # added and some important ones.
     doc = TaskDocument.from_logfile(p, ".log.gz").dict()
     assert doc["energy"] == pytest.approx(-4091.763)
-    assert doc["nsites"] == 2
+    assert doc["natoms"] == 2
     assert doc["charge"] == 0
     assert doc["spin_multiplicity"] == 3
     assert doc["nelectrons"] == 16
-    assert doc["molecule"] == doc["attributes"]["molecule_final"]
     assert "schemas" in doc["dir_name"]
     assert "gau_testopt.log.gz" in doc["logfile"]
-    assert doc.get("attributes", None) is not None
-    assert doc.get("metadata", None) is not None
+    assert doc.get("attributes") is not None
+    assert doc.get("metadata") is not None
     assert doc["metadata"]["success"] is True
     assert doc["attributes"]["molecule_initial"][0].coords == pytest.approx([0, 0, 0])
-    assert doc["attributes"]["molecule_final"][0].coords == pytest.approx(
-        [0.397382, 0.0, 0.0]
-    )
+    assert doc["molecule"][0].coords == pytest.approx([0.397382, 0.0, 0.0])
     assert doc["last_updated"] is not None
     assert doc["attributes"]["homo_energies"] == pytest.approx(
-        [-7.054007346511501, -11.618445074798501]
+        [-7.05400734, -11.61844507]
     )
     assert doc["attributes"]["lumo_energies"] == pytest.approx(
-        [4.2384453353880005, -3.9423854660440005]
+        [4.23844533, -3.94238546]
     )
     assert doc["attributes"]["homo_lumo_gaps"] == pytest.approx(
-        [11.292452681899501, 7.6760596087545006]
+        [11.29245268, 7.67605960]
     )
-    assert doc["attributes"]["min_homo_lumo_gap"] == pytest.approx(7.6760596087545006)
+    assert doc["attributes"]["min_homo_lumo_gap"] == pytest.approx(7.67605960)
 
     # Now we will try two possible extensions, but we will make sure that
     # it fails because the newest log file (.txt) is not valid
@@ -59,11 +55,13 @@ def test_cclib_taskdoc(test_dir):
     assert "Could not parse" in str(e.value)
 
     # Test a population analysis
-    doc = TaskDocument.from_logfile(p, ".out", analysis="MBO").dict()
+    doc = TaskDocument.from_logfile(p, "psi_test.out", analysis="MBO").dict()
     assert doc["attributes"]["mbo"] is not None
 
     # Let's try with two analysis (also check case-insensitivity)
-    doc = TaskDocument.from_logfile(p, ".out", analysis=["mbo", "density"]).dict()
+    doc = TaskDocument.from_logfile(
+        p, "psi_test.out", analysis=["mbo", "density"]
+    ).dict()
     assert doc["attributes"]["mbo"] is not None
     assert doc["attributes"]["density"] is not None
 
@@ -73,12 +71,12 @@ def test_cclib_taskdoc(test_dir):
 
     # Let's try a volumetric analysis
     # We'll gunzip the .cube.gz file because cclib can't read cube.gz files yet.
-    # Can remove the gzip part when https://github.com/cclib/cclib/issues/109 is closed.
+    # Can remove the gzip part when https://github.com/cclib/cclib/issues/108 is closed.
     with gzip.open(p / "psi_test.cube.gz", "r") as f_in, open(
         p / "psi_test.cube", "wb"
     ) as f_out:
         shutil.copyfileobj(f_in, f_out)
-    doc = TaskDocument.from_logfile(p, ".out", analysis=["Bader"]).dict()
+    doc = TaskDocument.from_logfile(p, "psi_test.out", analysis=["Bader"]).dict()
     os.remove(p / "psi_test.cube")
     assert doc["attributes"]["bader"] is not None
 
@@ -86,14 +84,19 @@ def test_cclib_taskdoc(test_dir):
     doc = TaskDocument.from_logfile(p, ".log", store_trajectory=True).dict()
     assert len(doc["attributes"]["trajectory"]) == 7
     assert doc["attributes"]["trajectory"][0] == doc["attributes"]["molecule_initial"]
-    assert doc["attributes"]["trajectory"][-1] == doc["attributes"]["molecule_final"]
+    assert doc["attributes"]["trajectory"][-1] == doc["molecule"]
 
     # Make sure additional fields can be stored
     doc = TaskDocument.from_logfile(p, ".log", additional_fields={"test": "hi"})
     assert doc.dict()["test"] == "hi"
 
+    # Test that the dict printing works
+    task = TaskDocument.from_logfile(p, "orca.out")
+    task.dict()
+
     # test document can be jsanitized
-    d = jsanitize(doc, strict=True, enum_values=True)
+    dct = jsanitize(doc, enum_values=True)
 
     # and decoded
-    MontyDecoder().process_decoded(d)
+    json_str = MontyDecoder().process_decoded(dct)
+    assert "builder_meta=EmmetMeta" in json_str
