@@ -1,10 +1,24 @@
+import logging
+import os
+import shutil
+import sys
+import tempfile
+from pathlib import Path
+from unittest import mock
+
 import pytest
+from fireworks import LaunchPad
+from jobflow import JobStore
+from jobflow.settings import JobflowSettings
+from maggma.stores import MemoryStore
+from monty.serialization import loadfn
+from pymatgen.core import Structure
+
+from atomate2.utils.log import initialize_logger
 
 
 @pytest.fixture(scope="session")
 def test_dir():
-    from pathlib import Path
-
     module_dir = Path(__file__).resolve().parent
     test_dir = module_dir / "test_data"
     return test_dir.resolve()
@@ -12,9 +26,6 @@ def test_dir():
 
 @pytest.fixture(scope="session")
 def log_to_stdout():
-    import logging
-    import sys
-
     # Set Logging
     root = logging.getLogger()
     root.setLevel(logging.DEBUG)
@@ -28,34 +39,27 @@ def log_to_stdout():
 
 @pytest.fixture(scope="session")
 def clean_dir(debug_mode):
-    import os
-    import shutil
-    import tempfile
-
     old_cwd = os.getcwd()
-    newpath = tempfile.mkdtemp()
-    os.chdir(newpath)
+    new_path = tempfile.mkdtemp()
+    os.chdir(new_path)
     yield
     if debug_mode:
-        print(f"Tests ran in {newpath}")
+        print(f"Tests ran in {new_path}")
     else:
         os.chdir(old_cwd)
-        shutil.rmtree(newpath)
+        shutil.rmtree(new_path)
 
 
-@pytest.fixture
+@pytest.fixture()
 def tmp_dir():
     """Same as clean_dir but is fresh for every test"""
-    import os
-    import shutil
-    import tempfile
 
     old_cwd = os.getcwd()
-    newpath = tempfile.mkdtemp()
-    os.chdir(newpath)
+    new_path = tempfile.mkdtemp()
+    os.chdir(new_path)
     yield
     os.chdir(old_cwd)
-    shutil.rmtree(newpath)
+    shutil.rmtree(new_path)
 
 
 @pytest.fixture(scope="session")
@@ -65,8 +69,6 @@ def debug_mode():
 
 @pytest.fixture(scope="session")
 def lpad(database, debug_mode):
-    from fireworks import LaunchPad
-
     lpad = LaunchPad(name=database)
     lpad.reset("", require_password=False)
     yield lpad
@@ -77,11 +79,8 @@ def lpad(database, debug_mode):
             lpad.db[coll].drop()
 
 
-@pytest.fixture(scope="function")
+@pytest.fixture()
 def memory_jobstore():
-    from jobflow import JobStore
-    from maggma.stores import MemoryStore
-
     store = JobStore(MemoryStore(), additional_stores={"data": MemoryStore()})
     store.connect()
 
@@ -89,27 +88,60 @@ def memory_jobstore():
 
 
 @pytest.fixture(scope="session", autouse=True)
-def log_to_stdout():
-    from atomate2.utils.log import initialize_logger
-
+def log_to_stdout_auto_use():
     initialize_logger()
 
 
-@pytest.fixture(scope="function")
+@pytest.fixture()
 def si_structure(test_dir):
-    from pymatgen.core import Structure
-
     return Structure.from_file(test_dir / "structures" / "Si.cif")
 
 
 @pytest.fixture(autouse=True)
 def mock_jobflow_settings(memory_jobstore):
     """Mock the jobflow settings to use our specific jobstore (with data store)."""
-    from unittest import mock
-
-    from jobflow.settings import JobflowSettings
 
     settings = JobflowSettings(JOB_STORE=memory_jobstore)
 
     with mock.patch("jobflow.SETTINGS", settings):
         yield
+
+
+@pytest.fixture(
+    params=[
+        "Si_227",
+        "Fe_229",
+        "S_58",
+        "Rb2P3_69",
+        "K2Au3_71",
+        "LaI3_63",
+        "KCeF4_123",
+        "RbO2_129",
+        "BaN2_15",
+        "TiNi_11",
+        "CaC2_2",
+        "KNO3_160",
+        "ZnO_186",
+    ],
+    ids=[
+        "F cubic",
+        "I cubic",
+        "P orth",
+        "I orth",
+        "F orth",
+        "C orth",
+        "P tet",
+        "I tet",
+        "C mono",
+        "P mono",
+        "tri",
+        "rhom",
+        "hex",
+    ],
+    scope="session",
+)
+def symmetry_structure(test_dir, request):
+    """The structures are copied from amset.
+    See https://github.com/hackingmaterials/amset/blob/main/tests/conftest.py
+    """
+    return loadfn(test_dir / "symmetry_structures" / f"{request.param}.json.gz")
