@@ -1,10 +1,24 @@
+from typing import TYPE_CHECKING
+
+import numpy as np
+from jobflow import JobStore, run_locally
+from maggma.stores.mongolike import MemoryStore
+from pymatgen.analysis.defects.generators import SubstitutionGenerator
+from pymatgen.core import Structure
+from pymatgen.io.vasp.outputs import WSWQ
+
+from atomate2.vasp.flows.defect import (
+    ConfigurationCoordinateMaker,
+    FormationEnergyMaker,
+    NonRadiativeMaker,
+)
+
+if TYPE_CHECKING:
+    from atomate2.common.schemas.defects import CCDDocument
+    from atomate2.vasp.schemas.defect import FiniteDifferenceDocument
+
+
 def test_ccd_maker(mock_vasp, clean_dir, test_dir):
-    from jobflow import run_locally
-    from pymatgen.core import Structure
-
-    from atomate2.vasp.flows.defect import ConfigurationCoordinateMaker
-    from atomate2.vasp.schemas.defect import CCDDocument
-
     # mapping from job name to directory containing test files
     # mapping from job name to directory containing test files
     ref_paths = {
@@ -38,11 +52,7 @@ def test_ccd_maker(mock_vasp, clean_dir, test_dir):
     flow = ccd_maker.make(si_defect, charge_state1=0, charge_state2=1)
 
     # run the flow and ensure that it finished running successfully
-    responses = run_locally(
-        flow,
-        create_folders=True,
-        ensure_success=True,
-    )
+    responses = run_locally(flow, create_folders=True, ensure_success=True)
 
     ccd: CCDDocument = responses[flow.jobs[-1].uuid][1].output
 
@@ -55,18 +65,6 @@ def test_ccd_maker(mock_vasp, clean_dir, test_dir):
 
 
 def test_nonrad_maker(mock_vasp, clean_dir, test_dir, monkeypatch):
-    import numpy as np
-    from jobflow import JobStore, run_locally
-    from maggma.stores.mongolike import MemoryStore
-    from pymatgen.core import Structure
-    from pymatgen.io.vasp.outputs import WSWQ
-
-    from atomate2.vasp.flows.defect import (
-        ConfigurationCoordinateMaker,
-        NonRadiativeMaker,
-    )
-    from atomate2.vasp.schemas.defect import FiniteDifferenceDocument
-
     # mapping from job name to directory containing test files
     ref_paths = {
         "relax q1": "Si_config_coord/relax_q1",
@@ -127,3 +125,70 @@ def test_nonrad_maker(mock_vasp, clean_dir, test_dir, monkeypatch):
         wswq_p.me_imag = np.array(wswq_p.me_imag)
         assert wswq_p.me_real.shape == (2, 4, 18, 18)
         assert wswq_p.me_imag.shape == (2, 4, 18, 18)
+
+
+def test_formation_energy_maker(mock_vasp, clean_dir, test_dir, monkeypatch):
+    from jobflow import SETTINGS, run_locally
+
+    # mapping from job name to directory containing test files
+    ref_paths = {
+        "bulk relax": "GaN_Mg_defect/bulk_relax",
+        "relax Mg_Ga-0 q=-2": "GaN_Mg_defect/relax_Mg_Ga-0_q=-2",
+        "relax Mg_Ga-0 q=-1": "GaN_Mg_defect/relax_Mg_Ga-0_q=-1",
+        "relax Mg_Ga-0 q=0": "GaN_Mg_defect/relax_Mg_Ga-0_q=0",
+        "relax Mg_Ga-0 q=1": "GaN_Mg_defect/relax_Mg_Ga-0_q=1",
+    }
+
+    fake_run_vasp_kwargs = {
+        k: {"incar_settings": ["ISIF"], "check_inputs": ["incar"]} for k in ref_paths
+    }
+
+    # automatically use fake VASP and write POTCAR.spec during the test
+    mock_vasp(ref_paths, fake_run_vasp_kwargs)
+
+    struct_GaN = Structure.from_file(test_dir / "structures" / "GaN.cif")
+    defects = list(
+        SubstitutionGenerator().get_defects(
+            structure=struct_GaN, substitution={"Ga": ["Mg"]}
+        )
+    )
+
+    maker = FormationEnergyMaker(
+        relax_radius="auto",
+        perturb=0.1,
+        collect_defect_entry_data=True,
+        validate_charge=False,
+    )
+    flow = maker.make(
+        defects[0],
+        supercell_matrix=[[2, 2, 0], [2, -2, 0], [0, 0, 1]],
+        defect_index=0,
+    )
+
+    # run the flow and ensure that it finished running successfully
+    _ = run_locally(
+        flow,
+        create_folders=True,
+        ensure_success=True,
+    )
+
+    def _check_plnr_locpot(name):
+        job = SETTINGS.JOB_STORE.query_one({"output.task_label": name})
+        plnr_locpot = job["output"]["calcs_reversed"][0]["output"]["locpot"]
+        assert set(plnr_locpot) == {"0", "1", "2"}
+
+    for k in ref_paths:
+        _check_plnr_locpot(k)
+
+    # make sure the the you can restart the calculation from prv
+    prv_dir = test_dir / "vasp/GaN_Mg_defect/bulk_relax/outputs"
+    flow2 = maker.make(
+        defects[0],
+        bulk_supercell_dir=prv_dir,
+        defect_index=0,
+    )
+    _ = run_locally(
+        flow2,
+        create_folders=True,
+        ensure_success=True,
+    )
