@@ -1,11 +1,25 @@
+import jobflow
+import numpy as np
+import pytest
+from emmet.core.tasks import TaskDoc
+from emmet.core.vasp.calculation import IonicStep, VaspObject
+from jobflow import run_locally
+from numpy.testing import assert_allclose
 from pytest import approx
+
+from atomate2.vasp.jobs.core import (
+    DielectricMaker,
+    HSERelaxMaker,
+    HSEStaticMaker,
+    MDMaker,
+    RelaxMaker,
+    StaticMaker,
+    TransmuterMaker,
+)
 
 
 def test_static_maker(mock_vasp, clean_dir, si_structure):
-    from jobflow import run_locally
-
-    from atomate2.vasp.jobs.core import StaticMaker
-    from atomate2.vasp.schemas.task import TaskDocument
+    jstore = jobflow.SETTINGS.JOB_STORE
 
     # mapping from job name to directory containing test files
     ref_paths = {"static": "Si_band_structure/static"}
@@ -19,23 +33,25 @@ def test_static_maker(mock_vasp, clean_dir, si_structure):
     mock_vasp(ref_paths, fake_run_vasp_kwargs)
 
     # generate job
-    job = StaticMaker().make(si_structure)
+    job = StaticMaker(task_document_kwargs={"store_volumetric_data": ("chgcar",)}).make(
+        si_structure
+    )
 
     # run the flow or job and ensure that it finished running successfully
     responses = run_locally(job, create_folders=True, ensure_success=True)
 
-    # validation the outputs of the job
+    # validate job outputs
     output1 = responses[job.uuid][1].output
-    assert isinstance(output1, TaskDocument)
+    assert isinstance(output1, TaskDoc)
     assert output1.output.energy == approx(-10.85037078)
+
+    with jstore.additional_stores["data"] as s:
+        doc = s.query_one({"job_uuid": job.uuid})
+        dd = doc["data"]
+        assert dd["@class"] == "Chgcar"
 
 
 def test_relax_maker(mock_vasp, clean_dir, si_structure):
-    from jobflow import run_locally
-
-    from atomate2.vasp.jobs.core import RelaxMaker
-    from atomate2.vasp.schemas.task import TaskDocument
-
     # mapping from job name to directory containing test files
     ref_paths = {"relax": "Si_double_relax/relax_1"}
 
@@ -51,20 +67,15 @@ def test_relax_maker(mock_vasp, clean_dir, si_structure):
     # run the flow or job and ensure that it finished running successfully
     responses = run_locally(job, create_folders=True, ensure_success=True)
 
-    # validation the outputs of the job
+    # validate job outputs
     output1 = responses[job.uuid][1].output
-    assert isinstance(output1, TaskDocument)
+    assert isinstance(output1, TaskDoc)
     assert output1.output.energy == approx(-10.85043620)
     assert len(output1.calcs_reversed[0].output.ionic_steps) == 1
     assert output1.input.parameters["NSW"] > 1
 
 
 def test_dielectric(mock_vasp, clean_dir, si_structure):
-    import numpy as np
-    from jobflow import run_locally
-
-    from atomate2.vasp.jobs.core import DielectricMaker
-
     # mapping from job name to directory containing test files
     ref_paths = {"dielectric": "Si_dielectric"}
 
@@ -83,12 +94,12 @@ def test_dielectric(mock_vasp, clean_dir, si_structure):
 
     # Additional validation on the outputs of the job
     output1 = responses[job.uuid][1].output
-    assert np.allclose(
+    assert_allclose(
         output1.calcs_reversed[0].output.epsilon_static,
         [[11.41539467, 0, 0], [0, 11.41539963, 0], [0, 0, 11.41539866]],
         atol=0.01,
     )
-    assert np.allclose(
+    assert_allclose(
         output1.calcs_reversed[0].output.epsilon_ionic,
         [[0, 0, 0], [0, 0, 0], [0, 0, 0]],
         atol=0.01,
@@ -96,11 +107,6 @@ def test_dielectric(mock_vasp, clean_dir, si_structure):
 
 
 def test_hse_relax(mock_vasp, clean_dir, si_structure):
-    from jobflow import run_locally
-
-    from atomate2.vasp.jobs.core import HSERelaxMaker
-    from atomate2.vasp.schemas.task import TaskDocument
-
     # mapping from job name to directory containing test files
     ref_paths = {"hse relax": "Si_hse_relax"}
 
@@ -119,18 +125,13 @@ def test_hse_relax(mock_vasp, clean_dir, si_structure):
 
     # validation on the output of the job
     output1 = responses[job.uuid][1].output
-    assert isinstance(output1, TaskDocument)
+    assert isinstance(output1, TaskDoc)
     assert output1.output.energy == approx(-12.5326576)
     assert len(output1.calcs_reversed[0].output.ionic_steps) == 3
     assert output1.input.parameters["NSW"] > 1
 
 
 def test_hse_static_maker(mock_vasp, clean_dir, si_structure):
-    from jobflow import run_locally
-
-    from atomate2.vasp.jobs.core import HSEStaticMaker
-    from atomate2.vasp.schemas.task import TaskDocument
-
     # mapping from job name to directory containing test files
     ref_paths = {"hse static": "Si_hse_band_structure/hse_static"}
 
@@ -149,19 +150,13 @@ def test_hse_static_maker(mock_vasp, clean_dir, si_structure):
     # run the flow or job and ensure that it finished running successfully
     responses = run_locally(job, create_folders=True, ensure_success=True)
 
-    # validation the outputs of the job
+    # validate job outputs
     output1 = responses[job.uuid][1].output
-    assert isinstance(output1, TaskDocument)
+    assert isinstance(output1, TaskDoc)
     assert output1.output.energy == approx(-12.52887403)
 
 
 def test_transmuter(mock_vasp, clean_dir, si_structure):
-    import numpy as np
-    from jobflow import run_locally
-
-    from atomate2.vasp.jobs.core import TransmuterMaker
-    from atomate2.vasp.schemas.task import TaskDocument
-
     # mapping from job name to directory containing test files
     ref_paths = {"transmuter": "Si_transmuter"}
 
@@ -183,7 +178,7 @@ def test_transmuter(mock_vasp, clean_dir, si_structure):
 
     # validate outputs
     output1 = responses[job.uuid][1].output
-    assert isinstance(output1, TaskDocument)
+    assert isinstance(output1, TaskDoc)
     assert output1.output.energy == approx(-21.33231747)
     assert output1.transformations["history"][0]["scaling_matrix"] == [
         [1, 0, 0],
@@ -196,13 +191,6 @@ def test_transmuter(mock_vasp, clean_dir, si_structure):
 
 
 def test_molecular_dynamics(mock_vasp, clean_dir, si_structure):
-    import pytest
-    from jobflow import run_locally
-
-    from atomate2.vasp.jobs.core import MDMaker
-    from atomate2.vasp.schemas.calculation import IonicStep, VaspObject
-    from atomate2.vasp.schemas.task import TaskDocument
-
     # mapping from job name to directory containing test files
     ref_paths = {"molecular dynamics": "Si_molecular_dynamics/molecular_dynamics"}
 
@@ -226,8 +214,8 @@ def test_molecular_dynamics(mock_vasp, clean_dir, si_structure):
 
     # generate job
     job = MDMaker().make(si_structure)
-    NSW = 3
-    job.maker.input_set_generator.user_incar_settings["NSW"] = NSW
+    nsw = 3
+    job.maker.input_set_generator.user_incar_settings["NSW"] = nsw
 
     # run the flow or job and ensure that it finished running successfully
     responses = run_locally(job, create_folders=True, ensure_success=True)
@@ -235,13 +223,13 @@ def test_molecular_dynamics(mock_vasp, clean_dir, si_structure):
     # validation on the output
 
     output1 = responses[job.uuid][1].output
-    assert isinstance(output1, TaskDocument)
+    assert isinstance(output1, TaskDoc)
     assert output1.output.energy == pytest.approx(-11.46520398)
 
     # check ionic steps stored as pymatgen Trajectory
     assert output1.calcs_reversed[0].output.ionic_steps is None
     traj = output1.vasp_objects[VaspObject.TRAJECTORY]
-    assert len(traj.frame_properties) == NSW
+    assert len(traj.frame_properties) == nsw
     # simply check a frame property can be converted to an IonicStep
     for frame in traj.frame_properties:
         IonicStep(**frame)
